moved more markdown rendering files to markdown dir
