+++ apps/web/src/lib/markdown/normalize-inline-math.ts
